+++ CHANGES.rst
@@ -1,3 +1,13 @@
+Release 1.2.0
+=========================================
+
+* **ENHANCEMENT:** Added support for the inclusion of scripts based on features used in the chart (#6).
+* **ENHANCEMENT:** Added ``dict`` support to ``options.series.labels.SeriesLabel.style`` and ``utility_classes.data_labels.DataLabel.style``.
+* **BUGFIX:** Fixed de-serialization error in ``options.series.data.geometric.GeometricZData`` which
+  prevented the population of ``.properties``.
+
+---------------------
+
 Release 1.1.1
 =========================================